2.4.6 - Rate limits
* added SetRatelimit Command
* added Toggle.Ratelimiting
* ratelimiting has been added this means that if enabled over a 10
second period if a user sends more than the Guild's Ratelimit their
messages will automatically be deleted and a DM will be sent telling
them that they are being rate limited. if a muted role is set up and
muteRepeatOffenders is enabled it S.A.I.L will now automatically mute
the user if they oversend by 3 messages and if an admin channel is set
up will send a message there else it will post the message in the
channel they abused the rate limit in. (This is not a slowmode and is
global across the whole guild.)