2.9.1.2
* fixed a bug that prevented event messages from appearing properly
* fixed a bug that would cause certain things to stall sail+
WhatsMyColour Command* changed how dms are shown to the bot owner
* lots of backend changes to improve performance* swapped newchar and
updatechar from alias to command name.+ added admin channel reporting
for the mute command.
* fixed a bug that caused the typing message to persist after commands
were run
* Hopefully fixed a bug that prevented reminders from sending
* fixed creator commands being hidden when running the Commands command
and Help command
* fixed a bug that would caused the editing of ccs when using an image
as a argument to ignore any mode arguments causing it to default to
replace and also adding said text to the command.
* fixed issues with animated Emojis by updating the lib