2.9.1.4
* updated usersettings to default to list mode.
* fixed a bug that would cause transferlevels to ignore users without a
profile created* fixed a bug that would cause admins or the server
owner from being able to manage a user's pixels* formatted the settings
and modules commands to show which modules are enabled and which are
disabled* Formatted the channels, modules and settings commands so that
they are consistant with other commands.
+ added mute messages that are sent to the admin channel of the server
when someone is muted+ added automodnote to the mute command
+ started implementation of javadocs
+ added system to prevent screenlog.0 overflow from occuring* Extended
error created when role heirachey stops sail from updating roles.
* fixed a bug that would cause the art pin list from being nuked,
* fixed the same bug for the dailymessagerequests list.
* extended modnote max char count for the list preview to 55
* Change the level up gif to an embedded gif rather than a file
* add aliases to delcc
* Rework mute command to be simpler
* Better startup validation
* Users that were in sail's muted list will automatically get re muted
on join event