adding unit test for the channel settings
added unit test for the channel setting
Only testing offline only code as i want it to run without it dying because the discord api doesn't allow 10000 request per second for unit test.

Also fixed a bug in the channel settings getDesc(Command) function where it checked to see if the desc was null and not if the command was null.

Also added a stray bracket that i mistakenly nuked out of existance when merging the code during the last pull request.