bug fixes
* Fixed a bug that caused the rulecode command to fail to respond or
update a user's roles.
* Fixed a bug that caused messages to work even if you didnt press space
after the command name leading to some command being impossible to run.
* Fixed serveral nullpointers within the modnote command. andi, what did
i say about testing your code?
* Fixed a bug in that the denyinvite catcher would throw an error on the
first message a user sent to a server.
+ Added a message thats sent when a muted user returns that is send to
the admin channel if set up.
* Fixed a bug that caused the char list to always show the author's list
rather than another user if specified.
* Fixed a bug where the rulecode role would automatically be handed out
to everyone if it existed rather than if the user had obtained the code
or not.